Shader precision fix.

+++ Assets/Kino/Bloom/Shader/Bloom.shader
@@ -68,7 +68,7 @@ Shader "Hidden/Kino/Bloom"
     half4 SafeHDR(half4 c) { return min(c, 65000); }
 
     // RGBM encoding/decoding
-    half4 EncodeHDR(half3 rgb)
+    half4 EncodeHDR(float3 rgb)
     {
     #if USE_RGBM
         rgb *= 1.0 / 8;
@@ -80,7 +80,7 @@ Shader "Hidden/Kino/Bloom"
     #endif
     }
 
-    half3 DecodeHDR(half4 rgba)
+    float3 DecodeHDR(half4 rgba)
     {
     #if USE_RGBM
         return rgba.rgb * rgba.a * 8;
